remove transformer_decoder_mp.py
